requirements.txt
amqp==5.1.1
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==5.1.1
Django==4.1.10
django-extensions==3.2.3
django-redis==5.3.0
djangorestframework==3.14.0
djongo==1.3.6
executing==1.2.0
idna==3.4
ipython==8.14.0
jedi==0.18.2
kombu==5.3.1
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.39
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pymongo==3.12.3
python-dateutil==2.8.2
pytz==2023.3
rabbitmq==0.2.0
redis==4.6.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.2.4
stack-data==0.6.2
traitlets==5.9.0
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
